Switch from TomDoc to YARD
This changes a lot of the documentation, but also changes the code a bit
to be able to document everything properly. The biggest changes are:

* No more setters in ContextIO::Account. These were private anyways, and
  code inside it was changed to adapt to this.
* ContextIO::Config now sets its attr_accessors explicitly.

The method visibility should now be specified both in the documentation
and in Ruby. The latter can be dropped if it makes for worse code inside
the library. Everything in the public API should be tagged @api public.
Everything not in the public API should be tagged @api private. This
means that you can easily find public and private API methods by running
yardoc with the --query option (--query="@api.text == 'public'" to only
get public API).

With this commit we have 100% documented. Let's try to keep it this way.